Update Input_Opt parameter in gc_allocate_all from INOUT to IN to reflect actual usage
This update makes the broadcast of Input_Opt prior to the call to gc_allocate_all
make more sense when reading the source code. gc_allocate_all actually requires
that all Input_Opt variables be broadcasted and then passed as input so that
Input_Opt%LUCX may be correctly used to set NAER, the number of aerosols, which
is different depending on if LUCX is true or false.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>