Set levels of lev and ilev dimensions properly in history_netcdf_mod.F90
We have fixed a minor issue in the GEOS-Chem "Classic" History
component.  In routine History_NetCDF_Define (located in module
History/history_netcdf_mod.F90), nLev was defined with Container%nZ
and nILev with Container%NZ_1. This would set e.g. nLev=72 and nILev=73.

But, if the collection were defined on level edges, then Container%NZ
already is equal to 73.  The above code would set nLev=73 and
nILev=74, causing an out-of-bounds error.

To fix this we now added an IF statement that defines nLev and
nILev properly regardless of whether or not the container is defined
on level edges.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>