Remove code for writing multiple time slices to file from hcoio_write_std_mod.F90
Christoph Keller added the capability to write multiple time slices to the
same file to HEMCO/Core/hcoio_writE_std_mod.F90. However, this messes up
writing the GEOS-Chem restart file because the code in restart_mod.F
depends on HEMCO routines. Using the multiple time slice update causes
GEOS-Chem to crash when writing the netCDF restart file with the error
"In NcGet_Var_Attr_C: time, NetCDF: Variable not found." As a quick fix,
we'll remove the code for multiple time slices for now, but in the future
we would like to remove the dependency of GEOS-Chem restart files on HEMCO
and utilize the History component instead.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>