Restore deleted code in history_netcdf_mod.F90
In the prior commit, the second call to Get_Var_DimID's (in
routine History_NetCDF_Write in History/history_netcdf_mod.F90)
was inadvertently deleted.  This caused all variables to be
written to netCDF as scalars, but not arrays.

We also needed to pass the iLevId argument to the second call
to NC_Var_Def.

With these two fixes, the proper netCDF otuput functionality has
been restored.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>