Restore use of prod/loss rates from GMI for linearized mesospheric chemistry in UCX-based simulations
Following the implemetation of the prod/loss rates from UCX in v11-02e,
it was discovered that we need to restore the use of prod/loss rates from
GMI in the mesosphere for UCX-based simulations. The prod/loss rates saved
out from a UCX simulation via KPP are only saved out on 59 levels (i.e. the
maximum number of levels in the chemistry grid). Values in levels 59-72
(~1 hPa to ~ 0.01 hPa according to our vertical grids wiki page) are set to
zero.

Chris Holmes wrote:

   I think it’s important to do some kind of mesospheric P/L in the UCX
   simulations. If we’re going to have a stratospheric chemistry capability
   in GEOS-Chem, we need to have a reasonable upper boundary condition (and
   zero P/L in the mesosphere is not reasonable). It seems to me that the
   easiest thing is to retain the GMI P/L in the mesosphere in UCX
   simulations.

To resolve this issue, we have added IF..ELSE statements to check if users
are running a UCX or tropchem simulation. UCX-based simulations will use
GMI P/L rates for linearized mesospheric chemistry and tropchem-based
simulations will use UCX P/L rates for linearized stratospheric chemistry.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>