Now activate EXPID for GEOS-Chem Classic simulations
The EXPID setting in HISTORY.rc allows you to specify the beginning
of the file path corresponding to each diagnostic collection.
For example:

   EXPID: ./GEOSChem
      Will create netCDF files whose names begin "GEOSChem",
      in the run directory.

   EXPID: ./OutputDir/GEOSChem
      Will create netCDF files whose names begin with "GEOSChem"
      in the OutputDir sub-folder of this run directory.

Also note: add an error trap to strip out the "./" from the
Container%FileExpId field in History/histcontainer_mod.F90.
Apparently the netCDF routines cannot write to file names with "./"
in them.  This fixes the issue.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>