Update GFED4 scale factors
This update was submitted by Katie Travis.
Message from Katie: "We scaled all emission factors for species emitted
on a per carbon basis explicitly in the file for transparency.  For
species not carried in the standard mechanism, such as C2H2 for example,
I took a guess based on previous work.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>