Replace MAX_PASV with number of passive species read from input.geos
Prior to this commit a hard-coded value stored in variable MAX_PASV was
used to allocate arrays related to passive species. This created arrays
that were unnecessarily large for the default of a single passive species,
or too small if a user wanted to use a large number of passive species.
Now the number of passive species is determined from input.geos and used
to allocate the arrays.

Also in this commit: rename subroutine Set_Input_Opt_Extra to
Set_Input_Opt_Advect so that there can be multiple Set_Input_Opt_x
subroutines with apt names.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>