In DIAG20, now divide LOx by mass of Ox instead of O3
It was pointed out that when converting the loss of Ox (LOx) from
molec/cm3/s to 1/m3/s, we should divide by the total mass of the
Ox family (in molec) instead of the mass of O3.

As a quick fix, we obtain the mass of Ox by summing over all of the
Ox constituent species in State_Chm%Species, and then multiplying by
the grid box volume.  Admittedly this is kind of a hack.  In the future,
we will look into obtaining the mass of Ox directly from KPP (or at
least having a way of letting KPP denote which members are in a family)
so that this can be implemented in a less-hardwired fashion.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>