Restore FINNv1.5 modules
Barron Henderson wrote:

   It looks like v1.6 2015 and 2016 are probably okay, but switching to v1.6 for
   all years would be a problem.
   So, we should hold off on the system-wide implementation for now.

  -Barron

We have therefore restored the HEMCO/Extensions/hcox_include.H and
HEMCO/Extensions/hcox_finn_mod.F90 files that are consistent with the FINN v1.5
data.   Also we have renamed the FINN v1.6 files with the extension *.FINNv16 so that
when it comes time to restore FINNv1.6, we can add these quickly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>