Add workaround for ifort 17.0.4 in Headers/state_diag_mod.F90
When using ifort 17.0.4 we were getting a seg fault in routine
INIT_STATE_DIAG (in Headers/state_diag_mod.F90) where we construct
the diagID variable for diagnostics at specific wavelengths.  In other
words:

  diagID  = 'AODDust' // TRIM( RadWL(3) ) // 'nm'

would cause seg fault.  This could be attributed to a compiler bug
in that particular version of ifort.  In any case, rewriting the
assignment to:

  TmpWL   = RadWL(3)
  diagID  = 'AODDust' // TRIM( TmpWL) ) // 'nm'

seems to get around this error.  We made this fix for all such instances
in routine INIT_STATE_DIAG.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>