Add ISOAAQ to AOD and PM2.5 diagnostics only if ComplexSOA is on
To avoid double-counting of SOA, we now add

  ISOAAQ = SOAGX + SOAMG + SOAIE + SOAME + LVOCOA + ISN10A

to the AOD diagnostics only when the complex SOA option is turned on.

Furthermore, for the PM2.5 diagnostic, we do the following:

(a) For simulations using Simple SOA only (Standard, Tropchem), then
    SOAS (but not ISOAAQ) will be added to the PM2.5 diagnostic;

(b) For simulations using Complex SOA only (complexSOA and
    complexSOA_SVPOA), then ISOAAQ (but not SOAS) will be added
    to the PM2.5 diagnostic;

(c) For the benchmark simulation (which turns on both simple and
    complex SOA simultaneously) SOAS but not ISOAAQ will be
    added to the PM2.5 diagnostic.  (This is to avoid double-counting.)

RESULTS:

This update will cause the Standard, Tropchem, and Benchmark
simulations to have differences in the PM2.5 diagnostic w/r/t the
prior commit (because SOAS and ISOAAQ together were added to the
PM2.5 diagnostic).

The complexSOA and complexSOA_SVPOA simulations will
have identical PM2.5 diagnostics w/r/t the prior commit (because
ISOAAQ was already being added to the PM2.5 diagnostic).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>