Suppress usage of BYNCY and LFR to avoid unnecessary GCHP error prints
ExtState%BYNCY and ExtState%LFR are used in the GEOS-5 environment but
not when running GCHP as a standalone model. However, both variable
DoUse fields were set to true in the most recent version of HEMCO brought
into GEOS-Chem v11-02f. This resulted in error tracebacks written to
the GCHP log by every thread when the import pointer to the variables
was not found, resulting in many unnecessary error messages when using
large numbers of cores.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>