Add comments to tagged_o3_mod.F
We have added comments to note that the sum of tagged O3 species #2 - #12
should match the total O3 species #1.

We also added a blurb reminding users to start with a zero-concentration
initial conditions file, and then to spin up for as many years as it
takes to get into steady-state.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>