Add extra error checks in HEMCO standalone module
When running a HEMCO standalone simulation, it was discovered that
if the species number ID's are mis-indexed, such as:

 # ID NAME   MW    MWEMIS MOLECRATIO  K0     CR     PKA
 1    CO     28.0   28.0   1.0        0.0    0.0    0.0
 3    SOAP  150.0  150.0   1.0        0.0    0.0    0.0
 4    NO     30.0   30.0   1.0        0.0    0.0    0.0

that this error is not trapped.  Instead, an out-of-bounds error
can overwrite species metadata such as molecular weights, leading
to incorrect results -- especially for those species that involve
unit conversions.

We have added several fixes to hco_standalone_mod.F90 to avoid
this situation.  We improve the algorithm for parsing the HEMCO
standalone species file.  In addition, we now throw errors when

(1) The ID# of the first species read from the HEMCO species file
    is not 1; and

(2) The ID# of the last species read from the HEMCO species file
    does not match the total number of species read in.

This should prevent any typos in the HEMCO standalone species file
from creating undetected errors.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>