Add P(H2O2) to default list of prod/loss families in chemical mechanisms
The P(H2O2) field is required as input in the aerosol-only simulations. To
simplify saving out new oxidant fields, we now include P(H2O2) by default
in the prod/loss families so that it is saved out when users turn on the
prod/loss diagnostic in either bpch or netCDF format.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>