Do not force L=1 by default in planeflight_mod.F
In the updates to the CH4 simulation added in v11-02f, code for forcing L=1
was added to routine PLANEFLIGHT when a PTYPE string was not recognized.
This meant users would always have to add new plane type strings to this
code to avoid setting the level to 1. Most users will use the planeflight
diagnostic for planeflight tracks, not surface observations so we have
modified this code to only set L=1 when a string for surface observation is
recognized (not the other way around). In addition, the code will stop with
a warning that level will be set to 1. If that is the users intention, then
they can comment out the call to GEOS_CHEM_STOP and rerun.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>