Do not compile microphysics packages (TOMAS, APM) when NC_DIAG=y
The TOMAS microphysics package needs to be compiled with BPCH_DIAG=y,
since (1) none of the existing TOMAS diagnostics have been ported to
netCDF and (2) some of the emissions pointers from HEMCO are defined
as diagnostics in hcoi_gc_diagn_mod.F90, which is only defined for
BPCH_DIAG=y.

In the interests of time, (so as not to delay the 12.0.0 release),
we have put an error trap in Makefile_header.mk that does not allow
you to compile TOMAS or APM when NC_DIAG=y.

We will wait for the TOMAS team to provide us with the relevant updates.

Also wrapped some TOMAS diagnostic code in #if defined(BPCH_DIAG) blocks,
which should make it easy to exclude these when compiling into
ESMs.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>