Extend UpdateAlarm in GCC History to intervals of months or years
We have modified the algorithm that computes HistContainer%UpdateAlarm
so that it follows what we do for HistContainer%FileWriteAlarm.
UpdateAlarm intervals are now permitted for e.g. N months or N years.

Also added a few more spaces the HistContainer_Print statement in
order to print the HistContainer%EpochJSec etc. fields properly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>