Bug fix: Turn on Synoz if Linoz is set to false in input.geos
There is a bug in strat_chem_mod.F90 where SYNOZ will never be
used. If LINOZ is turned off, then the code treats O3 like all
other species and uses the UCX (formerly GMI) prod/loss rates.
It looks like a logical LSYNOZ was added for GEOS-5 data
assimilation applications back in 2016, but LSYNOZ is not set
anywhere so it’s always False. This is easy to fix -- just set
LSYNOZ to true if LLINOZ is false in input.geos.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>