Add the MODEL_, MODEL_WRF pre-processor constants and use where not-ESMF specific.
This update adds two new pre-processor flags: MODEL_ and MODEL_WRF.

The former is used in many places along with the ESMF_ flag, with the intent of
replacing it where code is not ESMF-specific (does not use MAPL/ESMF) but instead
merely intended for GEOS-Chem to interface with an external model (e.g. accept
external date/time, pressure edges).

The latter is intended for future uses in the WRF-GC interface.

Signed-off-by: Haipeng Lin <linhaipeng@pku.edu.cn>