Disables UCXNETCDF for WRF-GC (MODEL_WRF)
This is a minor update that sets UCXNETCDF to FALSE for WRF-GC,
as NetCDF NOx coefficients are not pre-processed for the WRF grid.

This update might apply for any EXTERNAL_GRID, but conservatively
I chose to only apply it for MODEL_WRF for now. In fact UCX will STOP
if GRID4x5 and GRID2x25 is not defined and UCXNETCDF = TRUE.

Signed-off-by: Haipeng Lin <linhaipeng@pku.edu.cn>