Workaround for gfortran 8.2.0 internal compiler error
The GNU Fortran Compiler 8.2.0 throws an internal compiler error
when compiling the legacy module ocean_mercury_mod.F.  This error
goes away when manually reducing the optimization level to -O1.

Makefile_header.mk: Define a new variable IS_GNU_8 to denote when
we are using a gfortran 8.*.* version.

GeosCore/Makefile: Use IS_GNU_8 to test if we are using a gfortran 8.*.*
version.  If so, manually reduce the optimization level to -O1 for
module ocean_mercury_mod.F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>