Workaround for gfortran 8.2: Comment out debug variables in ucx_mod.F
In routines SETTLE_STRAT_AER and CALC_STRAT_AER, there are several
debugging variables that are defined as !$OMP+PRIVATE, but which
are initialized and accessed outside of parallel loops.  This caused
unit tests with gfortran 8.2.0 to fail (uninitialized variable error).

The simple solution is to just comment out all instances of these
variables.  These variables are only used to print debug information
to the log file and are not used elsewhere.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>