Fix compilation issue with Intel Fortran 18
In ifort 18.0.0 and higher, the -openmp switch (which activates OpenMP
parallelization) has been retired.  The new switch is named -qopenmp.

Now save the compiler version number to the COMPILER_VERSION variable
in Makefile_header.mk.  Use the value of COMPILER_VERSION to determine
whether to set the -qopenmp or -openmp switch.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>