Add a check for output grid overlap with original grid.
This update to regrid_a2a_mod adds a check to the reduced output grid when mapping in the E-W direction.
If the target grid has no overlap with the original grid, Xmap_R4R4/R8R8 will fail with an error
and halt the entire program. This update fixes this stop with a more graceful failure,
which returns missing values if output grid is determined to be size zero.

This is usually only a problem when running local grids like WRF-GC.

Signed-off-by: Haipeng Lin <linhaipeng@pku.edu.cn>