Simplify the test for netCDF-Fortran in Makefile_header.mk
We have rewritten the test that determines if the netCDF-Fortran
library is in the same folder or a different folder than the
netCDF-C library file.  The new test should be more robust and
forgiving of user error (i.e. specifying a GC_F_INCLUDE environment
variable when there is no separate netCDF-Fortran library.  This
new test also gracefully handles the special case of the
GEOS-Chem-Libraries package.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>