Remove unused Input_Opt variables related to binary diagnostics
Input_Opt%NDxx and Input_Opt%LDxx are not used for binary diagnostics
20, 23, 40, 48-51, and 63. However, Input_Opt variables were allocated
and initialized to zero for all of them and they were set in input_opt
to shadow variables for them that were never initialized. This can
cause a run-time error when running GEOS-Chem in GEOS-5. These Input_Opt
variables and the shadow variables in input_opt are now removed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>