Now register SSAlkAccum and SSAlkCoarse separately
The State_Chm%SSAlk array cannot be registered as a single array,
because it is has a 4th dimension of size 2, which is interpreted as per
species.

The solution is to register the two fields separately, as SSAlkAccum
and SSAlkCoarse.  This also requires a slight modification to the
HISTORY.rc files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>