Add NO_EXE=y option to Makefile to inhibit geos executable build.
This update to the Makefiles adds a new option NO_EXE=y to the Makefile,
for other models coupled to and driving GEOS-Chem from their routines.

When `NO_EXE=y` is enabled, the final `geos` executable is not built,
and `GeosCore` will build all prerequisite libraries leaving `libGeosCore.a`
in the `lib` folder instead for other models to link to in their compile
process.

When `NO_EXE=y` is enabled, the pre-processor directive `MODEL_` is
automatically set as GEOS-Chem must be operating under another model
when this option is enabled.

Signed-off-by: Haipeng Lin <linhaipeng@pku.edu.cn>