Add error trap stop run if any of the species names are blank
In routine Init_State_Chm (in Headers/state_chm_mod.F90), we have added
an error trap that will stop the run if any of the species names
in the species database (i.e. State_Chm%SpcData(N)%Info%Name) are
blank.  This has the potential to halt a run inadvertently.

We have seen an error of this type when the KPP mechanism was modified;
but were not able to substantiate where it came frmo.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>