Add fixes for NO_EXE option
The GEOS-Chem executable was not getting created and modules in
GeosCore were not compiled, even when not passing NO_EXE=y. To fix
this, we now define a variable EXE_NEEDED in Makefile_header.mk to
pass  to GeosCore/Makefile. The syntax in GeosCore/Makefile has
been fixed to use ifeq($(EXE_NEEDED),1) instead of ifdef NO_EXE
which will not work properly.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>