Remove FLEXGRID as an option for met field since it is now always used
We should still allow users to compile only with GEOS-FP or MERRA-2
met fields. FlexGrid will be used regardless of the met field chosen.

Use of the FLEXGRID C-preprocessor switch has also been removed in
other areas of the code where it served as a placeholder for Stage 2
code when we enable users to define a custom nested grid simulation
at run time. Stage 2 modifications will be brought into GEOS-Chem at
a later point. The FLEXGRID CPP switch is still defined in
Makefile_header.mk for use at that time.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>