Fix several issues in RAS convection code
We have made several fixes in various modules to correct issues from
the last commit.  The code now compiles but it dies with what appears
to be a division-by-zero error at line 1099 of ras_mod.F90.

In future commits, we might have to change the precision of the
RAS subroutine variables from GEOS-5.  Many of those are REAL*4, which
might cause roundoff errors.  We can also add better error traps
to prevent div-by-zero and similar errors.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>