Add a more robust fix for setting Intel compiler option -qopenmp
Christopher Holmes pointed out that the fix we added to GEOS-Chem 12.0.1
to select the "-qopenmp" flag instead of "-openmp" only worked for
Intel Fortran 18 but not for e.g. Intel Fortran 19.  We have now
added a more robust test to the Makefile_header.mk that should work
for any Intel compiler version number.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>