Re-enable AFCID emissions in hcox_dust_dead_mod.F
We have discovered that the bug in the AFCID fine PM2.5 dust emissions in
HEMCO/Extensions/hcox_dustdead_mod.F was caused by an array not being reset
to zero each time routine HCOX_DustDead_Run was called. This has now been
fixed and these emissions are ready for inclusion in GC 12.1.0.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>