Combine emissions and drydep budget diagnostics and refactor code
The emissions and dry deposition fluxes are applied to species
concentrations at the same time. For now, keep these two components
within the same tendencies diagnostics to mirror this.

The initial and final mass arrays are now stored in state_diag but
are not registered, and therefore are not output as diagnostics.
Using them allows better refactoring to greatly reduce code.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>