Make ModLoc a module parameter to remove need for init_diagnostics_mod
Init_Diagnostics_Mod included only setting a string variable for use
in error handling throughout the module. This can be accomplished by
simply making the module variable a parameter and defining it at the
top of the module.

With this update, init_diagnostics_mod is removed
and a compile error related to the error handling parameter is fixed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>