Remove obsolete CO2 sim code and reorder a loop for efficiency
In GeosCore/co2_mod.F, we have removed the code that was commented out
by Beata Bukosa and Christoph Keller.  The emissions for most tagged
CO2 species added into State_Chm%Species directly via HEMCO.  This update
has allowed us to remove much obsolete code in GeosCore/co2_mod.F.

Because the CO2 production from CO oxidation is the only species
that is still computed in co2_mod.F, we were able to reorder the
loop for efficiency.  The L-loop is now the outermost loop, which
is most efficient.  Also the entire loop is now blocked off with
an IF ( Input_Opt%LCHEMCO2 ) statement instead of having the IF
statement within the loop.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>