Add error handling for if # nc diagnostic tags does not match array size
Add error handling during diagnostic registry in state_diag_mod to check
that the total number of tags retrieved from the metadata string matches
the expected # of species as determined from the allocated array size.
If this error GEOS-Chem will exit with a useful error message indicating
the problem. If left unchecked, then it is possible to have netcdf
diagnostics pointing to garbage.

Also in this commit: simplify the budget computation code.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>