Add bug fix to routine COLLAPSE in hco_interp_mod.F90
There is a bug in the way HEMCO does the vertical interpolation for model
levels. Where the comments in ModelLev_Interpolate says it's lumping two
(four) levels, the code is actually lumping three (five) levels. This was
leading to an array-out-of-bounds error for G5_EDGE_NATIVE when attempting
to collapse fields on level edges from 73 levels to 78 levels. To resolve
the issue, we should use:

  ! Get maximum level to be used for pressure thickness calculations.
  TOPLEV = InLev1 + ( NLEV-1 )

  ! Thickness of output level
  THICK = EDG(1) - EDG(NLEV)

  ! Get level weights
  ALLOCATE(WGT(NLEV))
  WGT = 0.0
  DO I = 1, NLEV-1
     WGT(I) = ( EDG(I) - EDG(I+1) ) / THICK
  ENDDO

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>