Initial preps for Hg/POPs netCDF diags; Remove many calls to ERROR_STOP
We have started to make the initial preparations for the adding the
netCDF diagnostics to the Hg and POPs simulations.  We have moved
the call to Diagn_POPs out of the #ifdef BPCH_DIAG block in
GeosCore/hcoi_gc_diagn_mod.F90, similar to what we already do for
the Hg simulation.

Also removed calls to ERROR_STOP from many (but not all) remaining
locations throughout the code.  These are now replaced with calls
to GC_ERROR, which will allow us to pass the error code back out
to the main program level instead of immediately killing the run.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>