Separate HEMCO file write updates for standard model and GEOS-5 with ifdefs
This update is to test for zero differences between the standard model
before and after the merge with GEOS. File write updates cause difference
tests to fail due to differences in the output files. The standard model
updates not used in GEOS-5 are the additional metadata for COARDS
compliance and making TIME an 8-byte variable to avoid roundoffs. The
GEOS-5 only update is to accept writing multiple time slices
into the same file.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>