Pass pointers for all data types to Registry_Lookup
The simulation was crashing because it couldn't find 2D fields
in the registry for Chemistry State. To fix this, we just need
to add Ptr2d = Ptr2d to the call to Registry_Lookup in routine
History_AddItemToCollection (history_mod.F90). To avoid issues
like this in the future, we now pass pointers for all data types
to Registry_Lookup for Chemistry State, Meteorology State, and
Diagnostics State.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>