Add 'End' Option to frequency and duration settings in HISTORY.rc
In HISTORY.rc users can now set a collection's frequency and/or
duration to 'End'. When 'End' is specified, the date will be written
to the netCDF file at the end of the simulation. This is especially
useful for the Restart collection for saving fields needed for
subsequent GEOS-Chem runs. FileCloseYmd and FileCloseHms will be
computed as the amount of time between the start and end of the
simulation.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>