Bug fix - HEMCO phase condition
FlexGrid introduced a new HEMCO phase (0), meaning that GEOS-Chem
Classic HEMCO phases could be 0, 1, or 2. Emissions in GCC only
occur on HEMCO phase 2, so a condition in emissions_run was changed
from "If ( Phase == 1 ) Return" to "If ( Phase <= 1 ) Return".
However, this broke GCHP, which relies on a phase -1 for some
operations. Changing the condition to be "If (( Phase == 0 )
.or. ( Phase == 1 )) Return" fixes this.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>