Bug fix - wrong year being used for CFC emissions
The year for which CFC emissions are calculated is estimated
based on the Input_Opt%CFCYEAR variable. During initialization
this is set to 0 if we just want to use whatever the current
year is, but that doesn't work if in an ESMF environment. A fix
which was already in the code did not propagate correctly, so
instead a module variable is being used as a workaround.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>