Now zero POPs prodloss diagnostic arrays at top of CHEMPOPS routine
In GeosCore/pops_mod.F, we had to zero all of the fields of State_Diag
pertaining to the POPS diagnostic quantities.  This now makes sure that
we get the same output as for the bpch.

Also note: the bpch diagnostics have now been converted from kg to
kg/s so as to better match the netCDF diagnostic output. (This was
done mostly for convenience while implementing the netCDF diagnostics,
as the bpch diagnostics are soon to be removed from GEOS-Chem.)

Also renamed LossPOPPOCbyGasPhase to LossPOPPOCPObyGasPhase (and
similarly for OCPI, BCPO, BCPI) to reflect that this diagnostic
tracks the hydrophobic species.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>