Allow for SZA down to 98 degrees when getting photolysis rates
Eric Palmer (GMAO) wrote:
"I see that the code that calls the photolysis rates (in flexchem_mod.F90),
is only called when the solar zenith angle is < 90 degrees (COS(SZA) > 0.).
However, Fast-Jx allows for SZA down to 98 degrees, which is important in
the stratosphere-mesosphere where sunlight still illuminates at high
altitudes if the sun is below the horizon at the surface.

To allow for SZA down to 98 deg, I made the following one line change:

   IF ( State_Met%SUNCOSmid(I,J) > 0.e+0_fp ) THEN

to:

   IF ( State_Met%SUNCOSmid(I,J) > -0.13917e+0_fp ) THEN

(-0.13917e+0_fp = COS(98 degrees)

This looks like it has the desired effect, as species with sharp gradients
across the terminator (eg: Br, BrO), are now more "rounded" and look much
more like GMI (which also allows SZA down to 98 degrees). This also
eliminates the problem with the non-conservation of bromine family species
in the mesosphere due to transport across the terminator (ie, the sharp
edges of a square wave)."

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>