Add update to History_NetCDF_Define for Restart collection
An update has been added to routine History_NetCDF_Define so that
a netCDF file is defined but not created on the first timestep if
the collection is defined as instantaneous with frequency=duration.
This will avoid creation of a netCDF file filled with missing
values. The update was needed for the Restart collection to avoid
overwriting the restart file used to initialize the simulation, but
is useful for other collections where frequency=duration in
HISTORY.rc.

Also added a minor fix to update CollectionFileName after creating
the history container to avoid having "Not found" returned for
file name in the GEOS-Chem log file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>