Make GEOS HEMCO updates in lightning nox extension standard
This uses lightning diagnostic names instead of IDs to update diag arrays.

Also in this commit, remove extraneous MODEL_GEOS block for getting
pointer to ExtState%LFR. If LFR is used only in GEOS-5 then we should
go through and put all uses of it in blocks at a later time.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>