Add netCDF diagnostic output for the CO2 specialty simulation
Most of the CO2 specialty simulation diagnostics can be archived
via the HEMCO diagnostics facility.  The only diagnostic that could
not be archived in this way is the production of CO2 from CO
oxidation (formerly the ND04 #9 slot).

We have therefore added the ProdCO2fromCO diagnostic variables to
the State_Diag object in order to archive this quantity.  We have also
removed all of the slots of ND04 except for the chemical production
of CO2 from CO (aka category "CO2-SRCE", tracer "CO2che").

NOTE: The netCDF diagnostic ProdCO2fromCO now has units of kg/m2/s,
but the bpch "CO2che" diagnostic still has units of molec/cm2/s.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>