Disable writing multiple time slices to same file via HEMCO
This update causes an issues writing netcdf restart files via
restart_mod.F90. This same issue was previously identified in March 2018
and the same update was removed to avoid the problem. This time the
update will remain but is disabled by forcing IsOldFile to false. Once
the restart write is moved from HEMCO to HISTORY we can re-enable the
multiple time slice write feature in HEMCO.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>