Remove tagging from Hg2 diagnostics
According to Colin Thackray, the production and loss of Hg2 diagnostics
(in bpch, the PL-HG2-$ category) are only updated for the total Hg2
but not the tagged Hg2 species.  Therefore, we have removed the tagging
from the corresponding netCDF diagnostics.

Also, corrected a typo in gamap_mod.F where ND67 bpch diagnostics
for PREACC and PRECON had the wrong metadata units string.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>