Add netCDF diagnostics for some Hg emissions quantities
Several of the Hg emissions are archived via HEMCO manual diagnostics.
Therefore, for simplicity's sake, we have added these to the Mercury
collection of the netCDF diagnostics.

Added the Hg0 and Hg2/HgP emission quantities.  Still need to add
the ocean Hg etc. diagnostics.

NOTE: netCDF emissions are for total Hg species.  The tagged Hg
simulation is a bit of a mess to sort out, so we'll kick this can
down the road.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>