Now allow for exact matches in the routine Check_DiagList
The routine Check_DiagList (in Headers/diaglist_mod.F90) uses the
F90 INDEX command to test if a substring is found in a name.  The
test condition was INDEX( name, substring ) > 0.  This will return
TRUE if the substring is anywhere in the name.

In some instances it is necessary to force an exact match (i.e. to
avoid namespace confusion with other diagnostics during error checking),
so we have added an exact match option to Check_DiagList.  If called
with the Exact=.TRUE. optional argument, Check_DiagList will use
the test condition INDEX( name, substr ) == 1, which will return TRUE
the substring matches at the beginning of the name (in other words,
an exact match).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>