Extend CFC sfcmr fix in 12.0.3 for GCHP to WRF-GC
This update extends the fix to the bug in application of
CFC surface mixing ratios in GCHP implemented in GEOS-Chem 12.0.3
to WRF-GC by adding the MODEL_WRF pre-processor flag to the relevant
fixes in addition to MODEL_GEOS.

WRF-GC which uses a similar interface into G-C as GCHP, the Get_Year()
function does NOT return the correct simulation date (initialized as a dummy)
until the GEOS-Chem initialization routine is finished. Thus this fix
for GCHP will apply for WRF-GC and thus should be extended.