Assign names to all ND21 tracerdiag.dat entries to enable xbpch python read
ND21 tracers 3 and 60 are obsolete. However, entries are still written
for them in tracerinfo.dat but without short or long names. The lack of
name causes a problem reading the binary diagnostic data set into python
using xbpch. Using dummy names avoids this problem.

Ideally we would not longer write these obsolete diagnostics to file.
However, since binary diagnostics will be removed in this future we
are simply applying this quick fix to avoid the problem.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>