Removed some debug and bpch code from rrtmg_rad_transfer_mod.F
We have now removed the bpch code that reads the MODIS surface
albedo fields, as well as some debug printout.

NOTE: In this commit, the surface emissivity fields are still read
in as bpch, as the conversion of the files from bpch to netCDF
has not yet been completed.  The bpch reading code for surface
emissivity will be removed in a future commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>