Now only nullify 3-D fields of state_met when using GCHP/GEOS/WRF
We have added #ifdef statements so that if GEOS-Chem is being run
as GCHP, or coupled to NASA-GEOS or WRF, that we only nullify 3-D
fields instead of deallocating AND nullifying them.

Deallocating these fields seems to cause a seg fault.  This might be
because the fields of State_Met are attached to the ESMF/MAPL Import
state, which gets finalized later.  So you cannot deallocate the
State_Met fields while the ESMF/MAPL Imports are still not finalized.
That is our hypothesis, anyway.

In any case, the code in routine Cleanup_State_Met are now presented
in a more consistent way and should be easier to modify and maintain.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>