Restore PS_ID in SET_DRY_SURFACE_PRESSURE; Add fixes to Get_Met_Fields and Get_GC_Restart
To avoid issues in GCHP, the PS_ID argument is restored in routine
SET_DRY_SURFACE_PRESSURE in dao_mod.F. The call to this routine in
hcoi_gc_main_mod.F90 is now coded to pass PS_ID=2 to set PS2_DRY. PS1_DRY
will be set from the restart file if found or to PS2_DRY otherwise.

Also added fixes to move setting of State_Met and State_Chm fields in
Get_Met_Fields and Get_GC_Restart to outside of the IF ( am_I_Root )
statements.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>